dex: improve component layering / interface
