+++ crates/core/component/dex/src/component/circuit_breaker/value.rs
@@ -64,7 +64,7 @@ mod tests {
 
     use crate::component::position_manager::price_index::PositionByPriceIndex;
     use crate::component::router::HandleBatchSwaps as _;
-    use crate::component::{StateReadExt as _, StateWriteExt as _};
+    use crate::component::{InternalDexWrite, StateReadExt as _, SwapDataRead, SwapDataWrite};
     use crate::lp::plan::PositionWithdrawPlan;
     use crate::{
         component::{router::create_buy, tests::TempStorageExt},
