+++ crates/core/component/dex/src/component/mod.rs
@@ -19,8 +19,14 @@ pub use self::metrics::register_metrics;
 pub(crate) use arb::Arbitrage;
 pub(crate) use circuit_breaker::ExecutionCircuitBreaker;
 pub(crate) use circuit_breaker::ValueCircuitBreaker;
+pub(crate) use dex::InternalDexWrite;
 pub use dex::{Dex, StateReadExt, StateWriteExt};
-pub use position_manager::{PositionManager, PositionRead};
-pub use swap_manager::SwapManager;
+pub use position_manager::PositionManager;
+pub(crate) use swap_manager::SwapDataWrite;
+pub(crate) use swap_manager::SwapManager;
+
+// Read data from the Dex component;
+pub use position_manager::PositionRead;
+pub use swap_manager::SwapDataRead;
 #[cfg(test)]
 pub(crate) mod tests;
