+++ crates/core/component/dex/src/component/position_manager.rs
@@ -14,6 +14,7 @@ use tap::Tap;
 use tracing::instrument;
 
 use crate::component::{
+    dex::InternalDexWrite,
     dex::StateReadExt as _,
     position_manager::{
         base_liquidity_index::AssetByLiquidityIndex, inventory_index::PositionByInventoryIndex,
@@ -33,11 +34,10 @@ use crate::{event, state_key};
 use super::chandelier::Chandelier;
 
 const DYNAMIC_ASSET_LIMIT: usize = 10;
-const RECENTLY_ACCESSED_ASSET_LIMIT: usize = 10;
 
 mod base_liquidity_index;
-mod counter;
-mod inventory_index;
+pub(crate) mod counter;
+pub(crate) mod inventory_index;
 pub(crate) mod price_index;
 
 #[async_trait]
@@ -280,34 +280,6 @@ pub trait PositionManager: StateWrite + PositionRead {
         self.update_position(None, position).await?;
 
         Ok(())
-    }
-
-    /// Adds an asset ID to the list of recently accessed assets,
-    /// making it a candidate for the current block's arbitrage routing.
-    ///
-    /// This ensures that assets associated with recently active positions
-    /// will be eligible for arbitrage if mispriced positions are opened.
-    #[tracing::instrument(level = "debug", skip_all)]
-    fn add_recently_accessed_asset(
-        &mut self,
-        asset_id: asset::Id,
-        fixed_candidates: Arc<Vec<asset::Id>>,
-    ) {
-        let mut assets = self.recently_accessed_assets();
-
-        // Limit the number of recently accessed assets to prevent blowing
-        // up routing time.
-        if assets.len() >= RECENTLY_ACCESSED_ASSET_LIMIT {
-            return;
-        }
-
-        // If the asset is already in the fixed candidate list, don't insert it.
-        if fixed_candidates.contains(&asset_id) {
-            return;
-        }
-
-        assets.insert(asset_id);
-        self.object_put(state_key::recently_accessed_assets(), assets);
     }
 
     /// Record execution against an opened position.
